Updating listener
