--- src/AppBundle/EventListener/ImgUploadListener.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace AppBundle\EventListener;
-
-use Symfony\Component\HttpFoundation\File\UploadedFile;
-use Doctrine\ORM\Event\LifecycleEventArgs;
-use Doctrine\ORM\Event\PreUpdateEventArgs;
-use AppBundle\Entity\Post;
-use Symfony\Component\HttpFoundation\File\File;
-use AppBundle\Services\FileUploadManager;
-
-class ImgUploadListener
-{
-    private $uploader;
-
-    public function __construct(FileUploadManager $uploader)
-    {
-        $this->uploader = $uploader;
-    }
-
-    public function prePersist(LifecycleEventArgs $args)
-    {
-        $entity = $args->getEntity();
-
-        $this->uploadFile($entity);
-    }
-
-    public function preUpdate(PreUpdateEventArgs $args)
-    {
-        $entity = $args->getEntity();
-
-        $this->uploadFile($entity);
-    }
-
-    private function uploadFile($entity)
-    {
-        if (!$entity instanceof Post) {
-            return;
-        }
-
-        $file = $entity->getImage();
-
-        if (!$file instanceof UploadedFile) {
-            return;
-        }
-
-        $fileName = $this->uploader->upload($file);
-        $entity->setImage($fileName);
-    }
-
-    public function postLoad(LifecycleEventArgs $args)
-    {
-        $entity = $args->getEntity();
-
-        if (!$entity instanceof Post) {
-            return;
-        }
-
-        $fileName = $entity->getImage();
-
-        $entity->setImage(new File('images/'.$fileName));
-    }
-}
